+++ cmd/rabtap/broker_info_printer_test.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/jandelgado/rabtap/pkg"
+	rabtap "github.com/jandelgado/rabtap/pkg"
 	"github.com/jandelgado/rabtap/pkg/testcommon"
 	"github.com/stretchr/testify/assert"
 )
@@ -84,9 +84,9 @@ func ExampleBrokerInfoPrinter_Print() {
 	//     ├── amq.topic (exchange, type 'topic', [D])
 	//     ├── test-direct (exchange, type 'direct', [D|AD|I])
 	//     │   ├── direct-q1 (queue, key='direct-q1', running, [D])
-	//     │   │   ├── some_consumer (consumer user='guest', chan='172.17.0.1:40874 -> 172.17.0.2:5672 (1)')
+	//     │   │   ├── some_consumer (consumer user='guest', prefetch=0, chan='172.17.0.1:40874 -> 172.17.0.2:5672 (1)')
 	//     │   │   │   └── '172.17.0.1:40874 -> 172.17.0.2:5672' (connection client='https://github.com/streadway/amqp', host='172.17.0.2:5672', peer='172.17.0.1:40874')
-	//     │   │   └── another_consumer w/ faulty channel (consumer user='', chan='')
+	//     │   │   └── another_consumer w/ faulty channel (consumer user='', prefetch=0, chan='')
 	//     │   └── direct-q2 (queue, key='direct-q2', running, [D])
 	//     ├── test-fanout (exchange, type 'fanout', [D])
 	//     │   ├── fanout-q1 (queue, idle since 2017-05-25 19:14:32, [D])
@@ -98,6 +98,36 @@ func ExampleBrokerInfoPrinter_Print() {
 	//         ├── topic-q1 (queue, key='topic-q1', idle since 2017-05-25 19:14:17, [D|AD|EX])
 	//         └── topic-q2 (queue, key='topic-q2', idle since 2017-05-25 19:14:21, [D])
 
+}
+
+func ExampleBrokerInfoPrinter_printByConnection() {
+
+	mock := testcommon.NewRabbitAPIMock(testcommon.MockModeStd)
+	defer mock.Close()
+	client := rabtap.NewRabbitHTTPClient(mock.URL, &tls.Config{})
+
+	brokerInfoPrinter := NewBrokerInfoPrinter(
+		BrokerInfoPrinterConfig{
+			ShowStats:          false,
+			ShowByConnection:   true,
+			OmitEmptyExchanges: false,
+			NoColor:            true},
+	)
+	brokerInfo, err := client.BrokerInfo()
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	if err := brokerInfoPrinter.Print(brokerInfo, "http://rabbitmq/api", os.Stdout); err != nil {
+		log.Fatal(err)
+	}
+
+	// Output:
+	// http://rabbitmq/api (broker ver='3.6.9', mgmt ver='3.6.9', cluster='rabbit@08f57d1fe8ab')
+	// └── Vhost /
+	//     └── '172.17.0.1:40874 -> 172.17.0.2:5672' (connection client='https://github.com/streadway/amqp', host='172.17.0.2:5672', peer='172.17.0.1:40874')
+	//         └── some_consumer (consumer user='guest', prefetch=0, chan='172.17.0.1:40874 -> 172.17.0.2:5672 (1)')
+	//             └── direct-q1 (queue, running, [D])
 }
 
 func ExampleBrokerInfoPrinter_printWithQueueFilter() {