+++ cmd/rabtap/cmd_exchange.go
@@ -8,7 +8,7 @@ import (
 	"crypto/tls"
 	"os"
 
-	"github.com/jandelgado/rabtap/pkg"
+	rabtap "github.com/jandelgado/rabtap/pkg"
 	"github.com/streadway/amqp"
 )